# src/backend-web/docs/swagger.yaml
basePath: /
definitions:
  internal_user.AdminUpdateUserRequest:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        enum:
        - admin
        - user
        type: string
    type: object
  internal_user.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
  internal_user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_user.LoginResponse:
    properties:
      token:
        type: string
    type: object
  internal_user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: admin@backendweb.com
    name: Wander Suporte
  description: API de gerenciamento de usuários com autenticação JWT.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Web API
  version: "1.0"
paths:
  /admin/users:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todos os usuários ativos no sistema. Requer
        privilégios de administrador.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_user.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar todos os usuários
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Cria um usuário no sistema. Requer token de administrador.
      parameters:
      - description: Dados do novo usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_user.User'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Criar um novo usuário
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um usuário do sistema (Soft Delete). Requer privilégios
        de administrador.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deletar um usuário
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Permite alterar nome, e-mail ou cargo de um usuário via ID.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Campos a atualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.AdminUpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar qualquer usuário (Admin)
      tags:
      - Admin
  /login:
    post:
      consumes:
      - application/json
      description: Autentica o usuário e retorna um token JWT
      parameters:
      - description: Credenciais de login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Realizar login
      tags:
      - Public
securityDefinitions:
  BearerAuth:
    description: Digite 'Bearer ' seguido do seu token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"